Missing initial values for ODE's, look temporally in indicator before vertically
Before, we took values in 2010 and looked UP and DOWN cascade to populate max
and min values for indicators without fixed estimates.

However, if a value was provided in 2011+ then quite often (at least in terms
of numbbers diagnosed) we would be WAY off our estimate.

So, now the model looks along an indicator across all years and bases its MAX
value on a value LATER in time, the minimum is set to ZERO

However, then FindSense() prevents any non-sensical values being simulated so
everything is still a LITTLE constrained, but this is an artefact of the MODEL.